feat: export `proxifyArray` from library
